+++ apps/web-app/next.config.js
@@ -1,5 +1,11 @@
-const path = require('path');
+// @ts-check
+
+const { withSentryConfig } = require('@sentry/nextjs');
 const { i18n } = require('./next-i18next.config');
+
+// @ts-ignore
+const packageJson = require('./package');
+
 const NEXTJS_BUILD_TARGET = process.env.NEXTJS_BUILD_TARGET || 'server';
 const NEXTJS_IGNORE_ESLINT = process.env.NEXTJS_IGNORE_ESLINT === '1' || false;
 const isProd = process.env.NODE_ENV === 'production';
@@ -13,15 +19,17 @@ const tmModules = [
         // ie: '@react-google-maps/api'...
       ]
     : []),
-  // ESM only packages are not yet supported by NextJs
+  // ESM only packages are not yet supported by NextJs if you're not
+  // using experimental experimental esmExternals
+  // @link {https://nextjs.org/blog/next-11-1#es-modules-support|Blog 11.1.0}
+  // @link {https://github.com/vercel/next.js/discussions/27876|Discussion}
   // @link https://github.com/vercel/next.js/issues/23725
   // @link https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c
   ...[
     // ie: newer versions of https://github.com/sindresorhus packages
-    'ky',
   ],
 ];
-const withTM = require('next-transpile-modules')(tmModules, {
+const withNextTranspileModules = require('next-transpile-modules')(tmModules, {
   resolveSymlinks: true,
   debug: false,
 });
@@ -37,10 +45,6 @@ if (disableSourceMaps) {
     '[INFO]: Sourcemaps generation have been disabled through NEXT_DISABLE_SOURCEMAPS'
   );
 }
-
-const withBundleAnalyzer = require('@next/bundle-analyzer')({
-  enabled: process.env.ANALYZE === 'true',
-});
 
 // Example of setting up secure headers
 // @link https://github.com/jagaapple/next-secure-headers
@@ -63,64 +67,100 @@ const secureHeaders = createSecureHeaders({
   referrerPolicy: 'same-origin',
 });
 
-const config = withBundleAnalyzer(
-  withTM({
-    target: NEXTJS_BUILD_TARGET,
-    reactStrictMode: true,
-    webpack5: true,
-    productionBrowserSourceMaps: !disableSourceMaps,
-    i18n,
-    optimizeFonts: true,
+/**
+ * @type {import('next').NextConfig}
+ */
+const nextConfig = {
+  target: NEXTJS_BUILD_TARGET,
+  reactStrictMode: true,
+  // @ts-ignore
+  webpack5: true,
+  productionBrowserSourceMaps: !disableSourceMaps,
+  i18n,
+  optimizeFonts: true,
 
-    // @link https://nextjs.org/docs/basic-features/image-optimization
-    images: {
-      deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
-      imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
-      disableStaticImages: false,
-      // Allowed domains for next/image
-      domains: ['source.unsplash.com'],
-    },
+  httpAgentOptions: {
+    // @link https://nextjs.org/blog/next-11-1#builds--data-fetching
+    keepAlive: true,
+  },
 
-    eslint: {
-      ignoreDuringBuilds: NEXTJS_IGNORE_ESLINT,
-      dirs: ['src'],
-    },
+  experimental: {
+    // Prefer loading of ES Modules over CommonJS
+    // @link {https://nextjs.org/blog/next-11-1#es-modules-support|Blog 11.1.0}
+    // @link {https://github.com/vercel/next.js/discussions/27876|Discussion}
+    esmExternals: true,
+    // Experimental monorepo support
+    // @link {https://github.com/vercel/next.js/pull/22867|Original PR}
+    // @link {https://github.com/vercel/next.js/discussions/26420|Discussion}
+    externalDir: true,
+  },
 
-    async headers() {
-      return [{ source: '/(.*)', headers: secureHeaders }];
-    },
+  // @link https://nextjs.org/docs/basic-features/image-optimization
+  // @ts-ignore
+  images: {
+    deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
+    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
+    disableStaticImages: false,
+    loader: 'default',
+    // Allowed domains for next/image
+    domains: ['source.unsplash.com'],
+  },
 
-    webpack: function (config, { defaultLoaders, isServer }) {
-      // This extra config allows to use paths defined in tsconfig
-      // rather than next-transpile-modules.
-      // @link https://github.com/vercel/next.js/pull/13542
-      const resolvedBaseUrl = path.resolve(config.context, '../../');
-      config.module.rules = [
-        ...config.module.rules,
-        {
-          test: /\.(tsx|ts|js|jsx|json)$/,
-          include: [resolvedBaseUrl],
-          use: defaultLoaders.babel,
-          exclude: (excludePath) => {
-            return /node_modules/.test(excludePath);
-          },
-        },
-      ];
+  eslint: {
+    ignoreDuringBuilds: NEXTJS_IGNORE_ESLINT,
+    dirs: ['src'],
+  },
 
-      // A temp workaround for https://github.com/prisma/prisma/issues/6899#issuecomment-849126557
-      if (isServer) {
-        config.externals.push('_http_common');
-      }
+  async headers() {
+    return [{ source: '/(.*)', headers: secureHeaders }];
+  },
 
-      config.module.rules.push({
-        test: /\.svg$/,
-        issuer: /\.(js|ts)x?$/,
-        use: ['@svgr/webpack'],
-      });
+  // @ts-ignore
+  webpack: (config, { defaultLoaders, isServer }) => {
+    // A temp workaround for https://github.com/prisma/prisma/issues/6899#issuecomment-849126557
+    if (isServer) {
+      config.externals.push('_http_common');
+    }
 
-      return config;
-    },
-  })
-);
+    config.module.rules.push({
+      test: /\.svg$/,
+      issuer: /\.(js|ts)x?$/,
+      use: ['@svgr/webpack'],
+    });
+
+    return config;
+  },
+  env: {
+    APP_NAME: packageJson.name,
+    APP_VERSION: packageJson.version,
+    BUILD_TIME: new Date().getTime().toString(10),
+    SENTRY_RELEASE: process.env.SENTRY_RELEASE
+      ? process.env.SENTRY_RELEASE
+      : `${packageJson.name}@${packageJson.version}`,
+    NEXT_PUBLIC_SENTRY_DSN: process.env.SENTRY_DSN ?? '',
+  },
+  serverRuntimeConfig: {
+    // to bypass https://github.com/zeit/next.js/issues/8251
+    PROJECT_ROOT: __dirname,
+  },
+};
+
+let config = withNextTranspileModules(nextConfig);
+
+if (process.env.NEXT_DISABLE_SENTRY !== '1') {
+  config = withSentryConfig(config, {
+    dryRun:
+      process.env.NODE_ENV !== 'production' ||
+      process.env.NEXT_SENTRY_DRY_RUN === '1',
+  });
+}
+
+if (process.env.ANALYZE === 'true') {
+  // @ts-ignore
+  const withBundleAnalyzer = require('@next/bundle-analyzer')({
+    enabled: true,
+  });
+  config = withBundleAnalyzer(config);
+}
 
 module.exports = config;