+++ apps/web-app/src/backend/api/rest/poem-repository.ssr.ts
@@ -1,13 +1,13 @@
-import { PrismaClient } from '@your-org/db-main-prisma';
-import { InternalServerError, NotFound } from '@tsed/exceptions';
-import { Asserts, UnPromisify } from '@your-org/core-lib';
+import { PrismaClientDbMain } from '@your-org/db-main-prisma';
+import { InternalServerError } from '@tsed/exceptions';
+import { UnPromisify } from '@your-org/core-lib';
 
 export type GetPoems = UnPromisify<
   ReturnType<typeof PoemRepositorySsr['prototype']['getPoems']>
 >;
 
 export class PoemRepositorySsr {
-  constructor(private prisma: PrismaClient) {}
+  constructor(private prisma: PrismaClientDbMain) {}
 
   /**
    * @throws Error