+++ apps/web-app/src/features/home/home.config.ts
@@ -1,9 +1,9 @@
-import { CustomTypeOptions } from 'react-i18next';
+import { I18nNamespaces } from '@/core/i18n/i18n-namespaces.type';
 
-export const homeConfig = {
+export type HomeConfig = {
+  i18nNamespaces: Readonly<I18nNamespaces>;
+};
+export const homeConfig: HomeConfig = {
   /** Namespaces that should be loaded for this page */
-  i18nNamespaces: [
-    'common',
-    'home',
-  ] as (keyof CustomTypeOptions['resources'])[],
+  i18nNamespaces: ['common', 'home'],
 } as const;