+++ apps/web-app/src/features/home/pages/home.page.tsx
@@ -1,12 +1,11 @@
 import { NextSeo } from 'next-seo';
 import { MainLayout } from '@/components/layout/main-layout';
 import { Banner } from '@/components/banner';
-import { HeroSection } from './sections/hero-section';
-import { FeaturesSection } from './sections/features-section';
-import { CTASection } from './sections/cta-section';
+import { HeroBlock } from '../blocks/hero-block';
+import { FeaturesBlock } from '../blocks/features-block';
+import { CtaBlock } from '../blocks/cta-block';
 import { useTranslation } from 'next-i18next';
-import { homeConfig } from './home.config';
-import { CustomTypeOptions } from 'react-i18next';
+import { homeConfig } from '../home.config';
 import { AgenciesSection } from './sections/agencies-section';
 
 type Props = {
@@ -24,10 +23,10 @@ export const HomePage: React.FC<Props> = () => {
       />
       <MainLayout>
         <Banner />
-        <HeroSection />
+        <HeroBlock />
         <AgenciesSection />
-        <FeaturesSection />
-        <CTASection />
+        <FeaturesBlock />
+        <CtaBlock />
       </MainLayout>
     </>
   );