+++ apps/web-app/src/pages/home.tsx
@@ -1,21 +1,31 @@
-import { GetServerSideProps } from 'next';
-import { HomePage } from '../features/home/home.page';
+import { GetServerSideProps, InferGetServerSidePropsType } from 'next';
+import { HomePage } from '@/features/home/pages/home.page';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { BadRequest } from '@tsed/exceptions';
-import { homeConfig } from '../features/home/home.config';
+import { homeConfig } from '@/features/home/home.config';
 
-export default function HomeRoute() {
+type Props = {
+  /** Add HomeRoute props here */
+};
+
+export default function HomeRoute(
+  props: InferGetServerSidePropsType<typeof getServerSideProps>
+) {
   return <HomePage />;
 }
 
-export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
+export const getServerSideProps: GetServerSideProps<Props> = async (
+  context
+) => {
+  const { locale } = context;
   if (locale === undefined) {
     throw new BadRequest('locale is missing');
   }
   const { i18nNamespaces } = homeConfig;
   return {
     props: {
-      ...(await serverSideTranslations(locale, i18nNamespaces)),
+      // i18nNamespaces.slice() is needed here to get rid off readonly
+      ...(await serverSideTranslations(locale, i18nNamespaces.slice())),
     },
   };
 };