+++ packages/db-main-prisma/prisma/seeds/poem/poems.seed.ts
@@ -290,7 +290,7 @@ export const poemsSeed: Prisma.PoemCreateInput[] = [
     language: 'english',
     remove_digits: true,
     return_changed_case: true,
-    remove_duplicates: false,
+    remove_duplicates: true,
   });
 
   const poem: Prisma.PoemCreateInput = {